# fast-lane-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fast Lane Delivery API",
        "description": "API Documentation for Fast Lane Delivery and Guesthouse Booking service",
        "contact": {
            "email": "admin@fastlanedelivery.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/properties": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "List and search for properties",
                "description": "Returns a paginated list of verified properties. Can be filtered by quarter.",
                "operationId": "6e387a77d712ad025ae4dbbe58f9de35",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Property"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Properties"
                ],
                "summary": "Create a new property listing",
                "description": "Allows a host to create a new property. It will be unverified by default.",
                "operationId": "9f76887beafa8292ed82539c80bf512f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type",
                                    "description",
                                    "address",
                                    "quarter",
                                    "latitude",
                                    "longitude",
                                    "price_per_night"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Cozy Studio Apartment"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "guesthouse",
                                            "event_hall",
                                            "apartment"
                                        ],
                                        "example": "apartment"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A nice and cozy place to stay."
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Rue 1.234, Bastos"
                                    },
                                    "quarter": {
                                        "type": "string",
                                        "example": "Bastos"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3.8711
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 11.5179
                                    },
                                    "price_per_night": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 25000
                                    },
                                    "amenities": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Wi-Fi",
                                            "Air Conditioning"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Property created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Property"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "403": {
                        "description": "Forbidden, user is not a host"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/properties/{id}": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Get a specific property",
                "description": "Returns details of a specific property by ID.",
                "operationId": "fe8700ea91a142f83d73a651f84758fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the property",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Property"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Property not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Properties"
                ],
                "summary": "Update a property",
                "description": "Allows a host to update their property details.",
                "operationId": "4b96210ec02d19e06174634151963abf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the property",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Property"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Property updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Property"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "403": {
                        "description": "Forbidden, user is not the owner of the property"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Properties"
                ],
                "summary": "Delete a property",
                "description": "Allows a host to delete their property.",
                "operationId": "7f9b500db8d8101c62d2b387f1662ab7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the property",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Property deleted successfully"
                    },
                    "403": {
                        "description": "Forbidden, user is not the owner of the property"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Property": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "host_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "guesthouse",
                            "event_hall",
                            "apartment"
                        ],
                        "example": "guesthouse"
                    },
                    "name": {
                        "type": "string",
                        "example": "Villa Krystal"
                    },
                    "description": {
                        "type": "string",
                        "example": "A beautiful villa with a pool."
                    },
                    "address": {
                        "type": "string",
                        "example": "Rue 1.111, Nkolbisson"
                    },
                    "quarter": {
                        "type": "string",
                        "example": "Nkolbisson"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 3.8711
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 11.5179
                    },
                    "price_per_night": {
                        "type": "number",
                        "format": "float",
                        "example": 50000
                    },
                    "amenities": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Wi-Fi",
                            "Pool",
                            "Air Conditioning"
                        ]
                    },
                    "photos": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uri"
                        },
                        "example": [
                            "http://example.com/img1.jpg"
                        ]
                    },
                    "is_verified": {
                        "type": "boolean",
                        "readOnly": true,
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}